+++ src/main/java/roomescape/controller/ReservationController.java
@@ -2,14 +2,10 @@ package roomescape.controller;
 
 import jakarta.validation.Valid;
 import java.net.URI;
-import java.time.LocalDate;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.ReservationService;
 import roomescape.service.dto.AuthInfo;
@@ -38,7 +34,7 @@ public class ReservationController {
         return ResponseEntity.ok(reservationService.findMyReservations(authInfo));
     }
 
-    @PostMapping({"/reservations", "/reservations/waiting"})
+    @PostMapping({"/reservations"})
     public ResponseEntity<ReservationResponse> createReservation(
             @Valid @RequestBody ReservationCreateMemberRequest memberRequest, AuthInfo authInfo
     ) {
@@ -47,26 +43,11 @@ public class ReservationController {
         return ResponseEntity.created(URI.create("/reservations" + reservationResponse.id())).body(reservationResponse);
     }
 
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(
-            @Valid @RequestBody ReservationCreateRequest reservationCreateRequest) {
-        ReservationResponse reservationResponse = reservationService.save(reservationCreateRequest);
+    @PostMapping("/reservations/waiting")
+    public ResponseEntity<ReservationResponse> createWaiting(
+            @Valid @RequestBody ReservationCreateMemberRequest memberRequest, AuthInfo authInfo) {
+        ReservationCreateRequest reservationCreateRequest = ReservationCreateRequest.from(memberRequest, authInfo.id());
+        ReservationResponse reservationResponse = reservationService.wait(reservationCreateRequest);
         return ResponseEntity.created(URI.create("/reservations" + reservationResponse.id())).body(reservationResponse);
-    }
-
-    @GetMapping("/admin/reservations")
-    public ResponseEntity<ListResponse<ReservationResponse>> findBy(
-            @RequestParam(required = false) Long themeId,
-            @RequestParam(required = false) Long memberId,
-            @RequestParam(required = false) LocalDate dateFrom,
-            @RequestParam(required = false) LocalDate dateTo
-    ) {
-        return ResponseEntity.ok().body(reservationService.findBy(themeId, memberId, dateFrom, dateTo));
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteById(id);
-        return ResponseEntity.noContent().build();
     }
 }
