+++ src/main/resources/static/js/reservation-with-member.js
@@ -242,7 +242,7 @@ function requestDelete(id) {
     method: 'DELETE',
   };
 
-  return fetch(`${RESERVATION_API_ENDPOINT}/${id}`, requestOptions)
+  return fetch('/admin' + RESERVATION_API_ENDPOINT + `/${id}`, requestOptions)
       .then(response => {
         if (response.status !== 204) throw new Error('Delete failed');
       });
